+++ app/src/main/java/com/bearddev/cityguide/repositories/PlaceRepository.java
@@ -13,21 +13,21 @@ public class PlaceRepository
     public List<Place> getPlaces()
     {
         List<Place> places = new ArrayList<Place>(10);
-        places.add(new Place(1, "Wierzynek", "Pyszna polska restauracja", "Resto"));
-        places.add(new Place(2, "Szwagier", "Pyszna polska restauracja", "Resto"));
-        places.add(new Place(3, "Mac", "Pyszna polska restauracja", "Resto"));
-        places.add(new Place(4, "KFC", "Pyszna polska restauracja", "Resto"));
-        places.add(new Place(5, "Subway", "Pyszna polska restauracja", "Resto"));
-        places.add(new Place(6, "Sukiennice", "Pyszna polska restauracja", "Resto"));
-        places.add(new Place(7, "Grodza", "Pyszna polska restauracja", "Resto"));
-        places.add(new Place(8, "Zapiekanki na placu nowy", "Pyszna polska restauracja", "Resto"));
-        places.add(new Place(9, "Plac Inwalidow", "Pyszna polska restauracja", "Resto"));
-        places.add(new Place(10, "Pradnik Czerwony", "Pyszna polska restauracja", "Resto"));
+        places.add(new Place(1l, "Wierzynek", "Pyszna polska restauracja", "Resto"));
+        places.add(new Place(2l, "Szwagier", "Pyszna polska restauracja", "Resto"));
+        places.add(new Place(3l, "Mac", "Pyszna polska restauracja", "Resto"));
+        places.add(new Place(4l, "KFC", "Pyszna polska restauracja", "Resto"));
+        places.add(new Place(5l, "Subway", "Pyszna polska restauracja", "Resto"));
+        places.add(new Place(6l, "Sukiennice", "Pyszna polska restauracja", "Resto"));
+        places.add(new Place(7l, "Grodza", "Pyszna polska restauracja", "Resto"));
+        places.add(new Place(8l, "Zapiekanki na placu nowy", "Pyszna polska restauracja", "Resto"));
+        places.add(new Place(9l, "Plac Inwalidow", "Pyszna polska restauracja", "Resto"));
+        places.add(new Place(10l, "Pradnik Czerwony", "Pyszna polska restauracja", "Resto"));
 
         return places;
     }
 
-    public Place getPlaceByID(Integer placeID) {
-        return new Place(10, "Pradnik Czerwony", "Pyszna polska restauracja", "Resto");
+    public Place getPlaceByID(Long placeID) {
+        return new Place(10l, "Pradnik Czerwony", "Pyszna polska restauracja", "Resto");
     }
 }
